Moved util function to integration folder + allow for str
