+++ docs/source/en/tiktoken.md
@@ -43,7 +43,8 @@ The `tokenizer.model` file contains no information about additional tokens or pa
 Generate the `tokenizer.model` file with [tiktoken.get_encoding](https://github.com/openai/tiktoken/blob/63527649963def8c759b0f91f2eb69a40934e468/tiktoken/registry.py#L63) and then convert it to `tokenizer.json` with [`convert_tiktoken_to_fast`].
 
 ```py
-from transformers.tokenization_utils_fast import convert_tiktoken_to_fast
+
+from transformers.integrations.tiktoken import convert_tiktoken_to_fast
 from tiktoken import get_encoding
 
 # You can load your custom encoding or the one provided by OpenAI
@@ -51,8 +52,9 @@ encoding = get_encoding("gpt2")
 convert_tiktoken_to_fast(encoding, "config/save/dir")
 ```
 
-The resulting `tokenizer.json` file is saved to the specified directory and loaded with [`PreTrainedTokenizerFast`].
+The resulting `tokenizer.json` file is saved to the specified directory and can be loaded with [`PreTrainedTokenizerFast`].
 
 ```py
-PreTrainedTokenizerFast.from_pretrained("config/save/dir")
+tokenizer = PreTrainedTokenizerFast.from_pretrained("config/save/dir")
+```
